+++ hotqueue.py
@@ -4,6 +4,7 @@
 within your Python programs.
 """
 
+from contextlib import contextmanager
 from functools import wraps
 try:
     import cPickle as pickle
@@ -16,6 +17,9 @@ from redis import Redis
 __all__ = ['HotQueue']
 
 __version__ = '0.2.6'
+
+
+DEFAULT_BULK_SIZE = 500
 
 
 def key_for_name(name):
@@ -33,19 +37,19 @@ class HotQueue(object):
     :param name: name of the queue
     :param serializer: the class or module to serialize msgs with, must have
         methods or functions named ``dumps`` and ``loads``,
-        `pickle <http://docs.python.org/library/pickle.html>`_ will be used
-        if ``None`` is given
+        `pickle <http://docs.python.org/library/pickle.html>`_ is the default,
+        if ``None`` is given then messages will be passed to Redis without serialization
     :param kwargs: additional kwargs to pass to :class:`Redis`, most commonly
         :attr:`host`, :attr:`port`, :attr:`db`
     """
     
-    def __init__(self, name, serializer=None, **kwargs):
+    def __init__(self, name, serializer=pickle, **kwargs):
         self.name = name
-        if serializer is not None:
-            self.serializer = serializer
-        else:
-            self.serializer = pickle
+        self.serializer = serializer
         self.__redis = Redis(**kwargs)
+        self._bulk_mode = False
+        self._bulk_size = DEFAULT_BULK_SIZE
+        self._bulk_items = []
     
     def __len__(self):
         return self.__redis.llen(self.key)
@@ -102,7 +106,7 @@ class HotQueue(object):
                 msg = msg[1]
         else:
             msg = self.__redis.lpop(self.key)
-        if msg is not None:
+        if msg is not None and self.serializer:
             msg = self.serializer.loads(msg)
         return msg
     
@@ -113,9 +117,40 @@ class HotQueue(object):
         >>> queue.put("another message")
         """
         for msg in msgs:
-            msg = self.serializer.dumps(msg)
-            self.__redis.rpush(self.key, msg)
-    
+            if self.serializer:
+                msg = self.serializer.dumps(msg)
+            if self._bulk_mode:
+                self._bulk_items.append(msg)
+                if len(self._bulk_items) >= self._bulk_size:
+                    self._flush_bulk()
+            else:
+                self.__redis.rpush(self.key, msg)
+
+    @contextmanager
+    def bulk(self, bulk_size=None):
+        """Context manager for adding messages to the queue in bulk. Example:
+
+        >>> with queue.bulk(1000):
+                for counter in range(10000):
+                    queue.put('message %s' % counter)
+
+        :param bulk_size: the number of items to accumulate in memory before
+            flushing to Redis. 500 by default.
+        """
+        self._bulk_mode = True
+        self._bulk_size = bulk_size or DEFAULT_BULK_SIZE
+        yield
+        self._bulk_mode = False
+        self._flush_bulk()
+
+    def _flush_bulk(self):
+        """Flush the bulk queue, sending all the current messages
+        to Redis.
+        """
+        if self._bulk_items:
+            self.__redis.rpush(self.key, *self._bulk_items)
+            self._bulk_items = []
+
     def worker(self, *args, **kwargs):
         """Decorator for using a function as a queue worker. Example: